♻️ refactor(supabase): zmiana nazwy klienta suprabase na supabase
- zmiana nazwy pliku suprabaseClient.ts na supabaseClient.ts
- poprawki importów w api/contact/route.ts i api/health/route.ts
